+++ inst/tinytest/test-equivalence.R
@@ -2,7 +2,7 @@ source("helpers.R")
 requiet("poorman")
 requiet("emmeans")
 requiet("parameters")
-# exit_file("TODO")
+exit_file("TODO")
 
 mod <- lm(mpg ~ hp + factor(gear), data = mtcars)
 
@@ -58,13 +58,13 @@ N <- 100
 dat <- rbind(data.frame(y = rnorm(N), x = 0),
              data.frame(y = rnorm(N, mean = 0.3), x = 1))
 mod <- lm(y ~ x, data = dat)
-FUN <- function(model, ...) {
-    data.frame(term = "t-test", estimate = coef(model)[2])
+FUN <- function(x) {
+    data.frame(term = "t-test", estimate = coef(x)[2])
 }
 e1 <- tost(dat$y[dat$x == 0], dat$y[dat$x == 1], epsilon = .05)
 e2 <- hypotheses(
     mod,
-    FUN = FUN,
+    hypothesis = FUN,
     equivalence = c(-.05, .05),
     df = e1$parameter)
 expect_true(e1$tost.p.value > .5 && e1$tost.p.value < .9)
@@ -113,10 +113,12 @@ expect_inherits(x, "hypotheses")
 
 rm("mod")
 delta <- log(1.25)
-mod <<- lm(log(conc) ~ source + factor(percent), data = pigs)
+data(pigs, package = "emmeans")
+mod <- lm(log(conc) ~ source + factor(percent), data = pigs)
 rg <- ref_grid(mod)
 em <- emmeans(rg, "source", at = list(), df = Inf)
 pa <- pairs(em, df = Inf)
+
 mm <- predictions(
     mod,
     newdata = datagrid(grid_type = "balanced"),
@@ -125,16 +127,22 @@ mm <- predictions(
 
 e1 <- test(pa, delta = delta, adjust = "none", side = "nonsuperiority", df = Inf)
 e2 <- hypotheses(mm, equivalence = c(-delta, delta))
+e1 <- e1[order(e1$contrast),]
+e2 <- e2[order(e2$term),]
 expect_equivalent(e1$z.ratio, e2$statistic.nonsup, tol = 1e-6)
 expect_equivalent(e1$p.value, e2$p.value.nonsup, tol = 1e-6)
 
 e1 <- test(pa, delta = delta, adjust = "none", side = "noninferiority", df = Inf)
 e2 <- hypotheses(mm, equivalence = c(-delta, delta))
+e1 <- e1[order(e1$contrast),]
+e2 <- e2[order(e2$term),]
 expect_equivalent(e1$z.ratio, e2$statistic.noninf, tolerance = 1e-6)
 expect_equivalent(e1$p.value, e2$p.value.noninf)
 
 e1 <- test(pa, delta = delta, adjust = "none", df = Inf)
 e2 <- hypotheses(mm, equivalence = c(-delta, delta))
+e1 <- e1[order(e1$contrast),]
+e2 <- e2[order(e2$term),]
 expect_equivalent(e1$p.value, e2$p.value.equiv)