+++ NEWS.md
@@ -2,14 +2,26 @@
 
 ## Development
 
+Breaking changes:
+
+* `datagrid()` no longer includes the response variable by default when it is not explicitly specified by the user. Use the new `response` argument to include it.
+* `datagrid(grid_type="balanced")` returns factors ordered by level rather than by order of appearance in the original data.
+
 New:
 
 * `print.marginaleffects()` supports `style="tinytable"`, which returns a `tinytable` object. Call `print(avg_slopes(model))` to get a nice printed table in Quarto or Rmarkdown documents, via Typst, LaTeX or HTML.
+* print as HTML, Typst, or LaTeX format automatically via `tinytable`: `options(marginaleffects_print_style="tinytable")`
 * `hypothesis` argument accepts a function which takes a `marginaleffects` data frame and returns a transformed data frame with `term` and `estimate` columns.
+* `datagrid()` gets a `response` argument to control if the response variable is included or excluded from the grid-building process.
+* The `base::subset()` and `dplyr::filter()` functions can be called with a single argument in `newdata` to select a subset of rows from the dataset used to fit the model.
+  - Example: avg_comparisons(fit, variables = "treatment", newdata = subset(treatment == 1))`
+* Better warning for unsupported arguments.
 
 Bugfix:
 
 * Uninformative error when a custom `comparison` function returns `NA` in bayesian models. Thanks to @Sandhu-SS for report #1017.
+* `datagrid()` returns an object with full attributes when `by` is used. Thanks to @Sandhu-SS for report #1058.
+* `inferences(method="simulation")` with `inferences()` function. Thanks to @davidarmstrong for report #1054.
 
 ## 0.18.0